make it applicable for other types of mocks

+++ scripts/mock_tools/prepare_mocks_Y3.py
@@ -20,6 +20,7 @@ from LSS.globals import main
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--survey", help="e.g., Y1, DA2",default='DA2')
+parser.add_argument("--mockname", help="name of mocks", default='EZmock')
 parser.add_argument("--input_mockpath", help="full directory path to input mocks",default='')
 parser.add_argument("--input_mockfile", help="mock file name",default='')
 parser.add_argument("--output_fullpathfn", help="output mock file and full path",default='')
@@ -29,7 +30,7 @@ parser.add_argument("--zcol", help="name of column with redshift, including RSD"
 parser.add_argument("--ELGsplit", help="Are the ELGs split into LOP and VLO? If 'n', assuming all LOP",default='y')
 parser.add_argument("--ELGtpcol", help="column distinguishing the ELG type; assumed boolean with True being LOP",default='LOP')
 parser.add_argument("--ran_seed", help="seed for randoms; make sure this is different if running many in parallel",default=10)
-
+parser.add_argument("--nzmask", help="apply mask on galaxy number density n(z)",default='n')
 
 args = parser.parse_args()
 
@@ -43,25 +44,11 @@ elif args.tracer == 'BGS':
 tiletab = Table.read(f'/global/cfs/cdirs/desi/survey/catalogs/{args.survey}/LSS/tiles-{tile}.fits')
 
 
-tars = Table.read(args.input_mockpath+args.input_mockfile+".fits")
+data = Table.read(args.input_mockpath+args.input_mockfile)
 
 # Adding the WEIGHT column
-tars['WEIGHT'] = np.ones(tars['RA'].shape[0])
-
-# Conditions for NGC and SGC; AJR does not think this is necessary
-#condN = common.splitGC(tars)#(tars['RA'] > 85) & (tars['RA'] < 302)
-#condS = ~condN#(tars['RA'] < 85) | (tars['RA'] > 302)
-
-# Splitting the DataFrame; AJR: is there a reason for this here? 
-#tarsN = tars[condN]
-#tarsS = tars[condS]
-
-#tarsN["GALCAP"] = "N"
-#tarsS["GALCAP"] = "S"
-
-#data = vstack([tarsN, tarsS])
-data = Table(tars)#Table(data)
-del tars
+if 'WEIGHT' not in data.colnames:
+    data['WEIGHT'] = np.ones(data['RA'].shape[0])
 
 desitar = {'LRG':1, 'QSO': 4, 'ELG':34,'BGS': int(2**60)}
 priority = {'LRG':3200, 'QSO':3400, 'ELG':3100,'ELG_VOL':3000,'ELG_HIP':3200,'BGS':2100}
@@ -73,11 +60,12 @@ type_ = args.tracer
 data['DESI_TARGET'] = desitar[type_]
 data['PRIORITY_INIT'] = priority[type_]
 data['PRIORITY'] = priority[type_]
-if type_ == 'ELG' and args.ELGsplit == 'y':
-    sel_LOP = data[args.ELGtpcol] == 1
-    data['DESI_TARGET'][~sel_LOP] = 2+2**7
-    data['PRIORITY_INIT'][~sel_LOP] = 3000
-    data['PRIORITY'][~sel_LOP] = 3000
+if type_ == 'ELG':
+    if args.ELGsplit == 'y':
+        sel_LOP = data[args.ELGtpcol] == 1
+        data['DESI_TARGET'][~sel_LOP] = 2+2**7
+        data['PRIORITY_INIT'][~sel_LOP] = 3000
+        data['PRIORITY'][~sel_LOP] = 3000
     rans = rng.random(len(data))
     sel_HIP = rans < 0.1 #10% of ELG get promoted to HIP
     data['DESI_TARGET'][sel_HIP] += 2**6
@@ -88,7 +76,7 @@ if type_ == 'ELG' and args.ELGsplit == 'y':
 data['NUMOBS_MORE'] = numobs[type_]
 data['NUMOBS_INIT'] = numobs[type_]
 if type_ == 'QSO':
-    sel_highz = tars[args.zcol] > 2.1
+    sel_highz = data[args.zcol] > 2.1
     data['NUMOBS_MORE'][sel_highz] = 4
     data['NUMOBS_INIT'][sel_highz] = 4
     print('numobs counts',str(np.unique(data['NUMOBS_MORE'],return_counts=True)))
@@ -101,13 +89,12 @@ targets['TARGETID'] = (np.random.permutation(np.arange(1,n+1))+1e8*desitar[type_
 print(len(targets),' in Y5 area')
 selY3 = is_point_in_desi(tiletab,targets['RA'],targets['DEC'])
 targets = targets[selY3]
+if args.mockname == 'EZmock' and args.nzmask == 'y':
+    y3_nz_mask = (targets['STATUS']&2)>0   # match Y3 LRG/ELG/QSO n(z)
+    targets = targets[y3_nz_mask]
+
 print(len(targets),' in Y3 area')
 print('getting nobs and mask bits')
-
-if 'BRICKID' not in targets.colnames:
-	from desiutil import brick
-	tmp = brick.Bricks(bricksize=0.25)
-	targets['BRICKID'] = tmp.brickid(targets['RA'], targets['DEC'])
 
 def wrapper(bid_index):
 
@@ -128,26 +115,31 @@ def wrapper(bid_index):
 
     return data
 
-# Just some tricks to speed up things up
-bid_unique, bidcnts = np.unique(targets['BRICKID'], return_counts=True)
-bidcnts = np.insert(bidcnts, 0, 0)
-bidcnts = np.cumsum(bidcnts)
-bidorder = np.argsort(targets['BRICKID'])
-
-# start multiple worker processes
-with Pool(processes=int(args.nproc)) as pool: ##hay que poner un int para que funcione!
-    res = pool.map(wrapper, np.arange(len(bid_unique)))
-
-res = vstack(res)
-res.sort('idx')
-res.remove_column('idx')
-print('mask columns added')
-
-maskcols = ['NOBS_G','NOBS_R','NOBS_Z','MASKBITS']
-if np.array_equal(res['TARGETID'],targets['TARGETID']):
-    for col in maskcols:
-        targets[col] = res[col]
-    del res
+if 'MASKBITS' not in targets.colnames:
+    if 'BRICKID' not in targets.colnames:
+        from desiutil import brick
+        tmp = brick.Bricks(bricksize=0.25)
+        targets['BRICKID'] = tmp.brickid(targets['RA'], targets['DEC'])
+    # Just some tricks to speed up things
+    bid_unique, bidcnts = np.unique(targets['BRICKID'], return_counts=True)
+    bidcnts = np.insert(bidcnts, 0, 0)
+    bidcnts = np.cumsum(bidcnts)
+    bidorder = np.argsort(targets['BRICKID'])
+    
+    # start multiple worker processes
+    with Pool(processes=int(args.nproc)) as pool: ##hay que poner un int para que funcione!
+        res = pool.map(wrapper, np.arange(len(bid_unique)))
+    
+    res = vstack(res)
+    res.sort('idx')
+    res.remove_column('idx')
+    print('mask columns added')
+    
+    maskcols = ['NOBS_G','NOBS_R','NOBS_Z','MASKBITS']
+    if np.array_equal(res['TARGETID'],targets['TARGETID']):
+        for col in maskcols:
+            targets[col] = res[col]
+        del res
 
 mainp = main(tp = type_, specver = 'kibo')
 targets = common.cutphotmask(targets, bits=mainp.imbits)
